+++ test_mongodict.py
@@ -1,9 +1,13 @@
 # coding: utf-8
 
-import unittest
 import sys
+import unittest
+
 from collections import MutableMapping
+
 import pymongo
+
+from bson import Binary
 from mongodict import MongoDict
 
 
@@ -83,11 +87,15 @@ class TestMongoDict(unittest.TestCase):
         expected_keys = ['test-' + str(counter) for counter in range(10)]
         self.assertEqual(set(keys), set(expected_keys))
         self.assertEqual(set(my_dict.keys()), set(expected_keys))
+        my_dict['python'] = 10
+        expected_keys.append('python')
+        self.assertEqual(set(my_dict.keys()), set(expected_keys))
+
         results = []
         for key, value in my_dict.items():
             results.append((key, value))
         values = [x[1] for x in results]
-        expected_values = list(range(10))
+        expected_values = list(range(11))
         self.assertEqual(set(values), set(expected_values))
         self.assertEqual(set(my_dict.values()), set(expected_values))
 
@@ -101,12 +109,17 @@ class TestMongoDict(unittest.TestCase):
         for counter in range(10):
             self.collection.insert({'_id': 'test-' + str(counter),
                                     'value': counter})
-        my_dict = MongoDict()
-        my_dict.clear() # should use collections' drop method
+        my_dict = MongoDict(**self.config)
+        index_count_before = len(self.collection.index_information())
+        my_dict.clear()
+        index_count_after = len(self.collection.index_information())
         self.assertEqual(self.collection.find().count(), 0)
 
+        # MongoDict.clear() should not delete indexes
+        self.assertEqual(index_count_before, index_count_after)
+
     def test_should_be_possible_to_assign_new_values_to_existing_keys(self):
-        my_dict = MongoDict()
+        my_dict = MongoDict(**self.config)
         my_dict['python'] = 'rules'
         my_dict['python'] = 42
         self.assertNotEqual(my_dict['python'], 'rules')
@@ -114,23 +127,56 @@ class TestMongoDict(unittest.TestCase):
 
 
     def test_non_unicode_strings(self):
-        my_dict = MongoDict()
-        if sys.version < '3':
-            string_1 = unicode('Álvaro Justen'.decode('utf-8'))\
-                .encode('iso-8859-15')
-        else:
-            string_1 = 'Álvaro Justen'.encode('iso-8859-15')
+        my_dict = MongoDict(**self.config)
+        if sys.version >= '3':
+            return # pymongo on Python 3 does not have this problem
 
-        with self.assertRaises(UnicodeError):
+        string_1 = 'Álvaro Justen'.decode('utf8').encode('iso-8859-15')
+        with self.assertRaises(pymongo.errors.InvalidStringData):
             my_dict[string_1] = 123
-        with self.assertRaises(UnicodeError):
+        with self.assertRaises(pymongo.errors.InvalidStringData):
             temp = my_dict[string_1]
-        with self.assertRaises(UnicodeError):
+        with self.assertRaises(pymongo.errors.InvalidStringData):
             my_dict['python'] = string_1
+        with self.assertRaises(pymongo.errors.InvalidStringData):
+            string_1 in my_dict
+        with self.assertRaises(pymongo.errors.InvalidStringData):
+            del my_dict[string_1]
 
     def test_deletion_of_MongoDict_object(self):
-        my_dict = MongoDict(safe=False)
+        config = self.config.copy()
+        config['safe'] = False
+        my_dict = MongoDict(**config)
         for i in range(1000):
             my_dict['testing_' + str(i)] = i
         del my_dict
         self.assertEqual(self.collection.find().count(), 1000)
+
+    def test_keys_method_should_not_raises_exception_if_more_than_16MB(self):
+        my_dict = MongoDict(**self.config)
+        key_template = ('python' * 100) + '{}'
+        key_byte_count = 0
+        key_count = 0
+        keys = set()
+        while key_byte_count < 20 * 1024 * 1024: # 20MB > 16MB
+            new_key = key_template.format(key_count)
+            my_dict[new_key] = 'some value'
+            key_byte_count += len(new_key)
+            key_count += 1
+            keys.add(new_key)
+        dict_keys = my_dict.keys()
+        self.assertEquals(len(keys), len(dict_keys))
+        self.assertTrue(keys == set(dict_keys))
+        # do not use assertEqual here! The content is too big
+
+    def test_Binary(self):
+        my_dict = MongoDict(**self.config)
+        if sys.version < '3':
+            invalid_utf8 = 'á'.decode('utf-8').encode('iso-8859-15')
+        else:
+            invalid_utf8 = 'á'.encode('iso-8859-15')
+        binary_data = Binary(invalid_utf8)
+        my_dict['test'] = binary_data
+        result = my_dict['test']
+        self.assertIn(type(result), (bytes, Binary))
+        self.assertEqual(Binary(my_dict['test']), binary_data)